Improved documentation
The documentation is still lacking in many part of the codebase. The
setup.py didn't have the information to populate the readthedocs
page for the pypi package.

Two examples were added to help understand how this library can
be used as a library. The documentation on read the docs
does provide a documentation for the command line tool.

Some documentation was added in the files:

- odoo_tools/api/objects.py
- odoo_tools/api/environment.py
- odoo_tools/modules/search.py
- odoo_tools/env.py

Some of those files had improperly formatted documentation.