Upgrade PyOpenSSL to higher than 21 because of issues with cryptography
It makes more sense to increase the PyOpenSSL version than to downgrade
cryptography to an older version. This way both PyOpenSSL can be
upgraded to a newer version if not already installed.

PyOpenSSL drives required version of cryptography so if something
breaks later, it will be possible to either upgrade PyOpenSSL
or freeze PyOpenSSL if its incompatible with Odoo code itself.