Added OdooApplication and Plugins
The OdooApplication serves as a modular extension of Odoo. The current
implementation is still quite tied to Odoo as it's difficult to split
Odoo in isolated components.

Currently there are at least a few components that can be divided
relatively cleanly.

The goal is to provide a common API for various version of Odoo while
having an API that is clean and safe to use. The big WSGI application
can be split in multiple trait that provide certain capabilities like
handling session, routing requests etc.

Static Assets Middleware
========================

This can live in its own middleware outside of Odoo. In practice it only
need to know where the modules are located to be able to dispatch the
static assets. When the static assets is split from odoo, it's possible
to create a common interface to build static assets middleware up for
implementation.

Session Store
=============

This is an important component of Odoo that can be implemented as a
special mixin for the WSGI App. As it is, the Session Store can
be implemented using a common interface that make a session store
and provide a session. The rest of the API should be common for every
session store. For backward compatibility it also provides an interface
for session_gc even if some session store may not need it. This could
be removed in the future to simplify the API.

Dispatchers & Routers
=====================

These components are build to be created once unlike how Odoo does.
There's no need to create dispatchers on every requests if they can
handle many requests. This could be improved in the future to prevent
mutating the request object. Unfortunately, Odoo requires a lot
of things to be modified when matching a route. Currently, it's trying
to be as close as Odoo as possible to work but having a cleaner API in
the future would be better and safer.

WSGI Mixins
===========

The rest of the spaghetti monster that is in the Root object of Odoo.