Limit cryptography requirements to < 39
Recently, cryptography deprecated some x509 imports. It renders
the version of cryptography below 39 and above 39 incompatible.

Urllib3 under 1.26 do not fix a version to the cryptography module
so using a version of urllib3 under 1.26 will pull cryptography 39
or above despite being incompatible.

A temporary fix would be to fix cryptography to lower than 39 and
eventually force version of urllib3 above or equal to 1.26 to
prevent pulling a wrong version of cryptography.