Refactored slightly the environment variable parser.
The environment variable parser now works with EnvironmentVariable
properties. Each properties can be easily documented with their own
respective docstrings. In the past, each properties were declared
into the __init__ of the class. Now those properties are members
of the class.

There are a couple of other properties that are added to be able
to Store the value in the owner class and to be able to map values
from environment variables to python and back into environment
variables.

Some bugs were fixed as a result to parse properly the environment
variables.

Documentation has been improved to match some of the changes made. A new
section called references was added to document python modules without
actual specific documentation like user guide/dev guides.